// Assets-Uncapped/front-portal-develop/src/services/api/connections/models/PlaidLinkErrorPayload.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-connections
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: Unknown version
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    PlaidLinkError,
    PlaidLinkErrorFromJSON,
    PlaidLinkErrorFromJSONTyped,
    PlaidLinkErrorToJSON,
} from './PlaidLinkError';
import {
    PlaidLinkOnExitMetadata,
    PlaidLinkOnExitMetadataFromJSON,
    PlaidLinkOnExitMetadataFromJSONTyped,
    PlaidLinkOnExitMetadataToJSON,
} from './PlaidLinkOnExitMetadata';

/**
 * 
 * @export
 * @interface PlaidLinkErrorPayload
 */
export interface PlaidLinkErrorPayload {
    /**
     * 
     * @type {PlaidLinkError}
     * @memberof PlaidLinkErrorPayload
     */
    error?: PlaidLinkError;
    /**
     * 
     * @type {PlaidLinkOnExitMetadata}
     * @memberof PlaidLinkErrorPayload
     */
    metadata?: PlaidLinkOnExitMetadata;
}

export function PlaidLinkErrorPayloadFromJSON(json: any): PlaidLinkErrorPayload {
    return PlaidLinkErrorPayloadFromJSONTyped(json, false);
}

export function PlaidLinkErrorPayloadFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlaidLinkErrorPayload {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'error': !exists(json, 'error') ? undefined : PlaidLinkErrorFromJSON(json['error']),
        'metadata': !exists(json, 'metadata') ? undefined : PlaidLinkOnExitMetadataFromJSON(json['metadata']),
    };
}

export function PlaidLinkErrorPayloadToJSON(value?: PlaidLinkErrorPayload | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'error': PlaidLinkErrorToJSON(value.error),
        'metadata': PlaidLinkOnExitMetadataToJSON(value.metadata),
    };
}