// Assets-Uncapped/front-portal-develop/src/services/api/connections/models/PlaidUserEvent.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-connections
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: Unknown version
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    PlaidUserEventMetadata,
    PlaidUserEventMetadataFromJSON,
    PlaidUserEventMetadataFromJSONTyped,
    PlaidUserEventMetadataToJSON,
} from './PlaidUserEventMetadata';

/**
 * 
 * @export
 * @interface PlaidUserEvent
 */
export interface PlaidUserEvent {
    /**
     * 
     * @type {string}
     * @memberof PlaidUserEvent
     */
    eventName?: string;
    /**
     * 
     * @type {PlaidUserEventMetadata}
     * @memberof PlaidUserEvent
     */
    metadata?: PlaidUserEventMetadata;
}

export function PlaidUserEventFromJSON(json: any): PlaidUserEvent {
    return PlaidUserEventFromJSONTyped(json, false);
}

export function PlaidUserEventFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlaidUserEvent {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'eventName': !exists(json, 'eventName') ? undefined : json['eventName'],
        'metadata': !exists(json, 'metadata') ? undefined : PlaidUserEventMetadataFromJSON(json['metadata']),
    };
}

export function PlaidUserEventToJSON(value?: PlaidUserEvent | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'eventName': value.eventName,
        'metadata': PlaidUserEventMetadataToJSON(value.metadata),
    };
}